requirements.txt
azure-cognitiveservices-speech
requests
python-dotenv
pyttsx3